requirements.txt
aiodns==1.1.1
aiohttp==3.7.4
alembic==1.4.3
amqp==2.6.1
appdirs==1.4.4
asgiref==3.2.10
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==20.2.0
Automat==20.2.0
backoff==1.10.0
backports.entry-points-selectable==1.1.0
backtrader==1.9.75.123
banal==1.0.1
bcrypt==3.2.0
beautifulsoup4==4.8.2
behave==1.2.5
billiard==3.6.4.0
black==19.10b0
branca==0.4.2
bs4==0.0.1
cached-property==1.5.2
cachetools==4.1.1
cattrs==1.0.0
ccxt==1.53.25
celery==4.4.1
certifi==2019.11.28
cffi==1.14.2
cfgv==3.2.0
chardet==3.0.4
charset-normalizer==2.0.6
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.3
constantly==15.1.0
croniter==1.0.15
cryptography==3.2
cssselect==1.1.0
cycler==0.10.0
dataset==0.8.0
decorator==4.4.2
distlib==0.3.1
distro==1.6.0
Django==3.1.1
dnspython==1.16.0
enum-compat==0.0.3
et-xmlfile==1.0.1
eventlet==0.17.4
fake-useragent==0.1.11
Faker==4.1.2
fastquant==0.1.5.0
filelock==3.0.12
flower==0.9.5
folium==0.12.1
forex-python==1.5
future==0.18.2
geocell==0.9.0
geocoder==1.38.1
gevent==21.1.2
glob2==0.5
google-api-core==1.22.4
google-api-python-client==2.14.0
google-auth==1.22.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.52.0
googlemaps==4.4.5
greenlet==1.1.0
haversine==2.3.1
httplib2==0.19.1
humanize==3.1.0
hyperlink==20.0.1
identify==1.5.6
idna==2.8
idna-ssl==1.1.0
image==1.5.16
importlib-metadata==4.6.1
importlib-resources==3.0.0
incremental==17.5.0
iniconfig==1.0.1
itemadapter==0.2.0
itemloaders==1.0.4
jdcal==1.4.1
Jinja2==3.0.1
jmespath==0.10.0
joblib==0.17.0
jsonmerge==1.7.0
jsonpickle==0.9.5
jsonschema==3.2.0
kiwisolver==1.2.0
kombu==4.6.11
lxml==4.6.2
Mako==1.0.6
MarkupSafe==2.0.1
matplotlib==3.1.2
monotonic==1.5
more-itertools==8.5.0
multidict==4.7.6
multitasking==0.0.9
networkx==2.4
nltk==3.5
nodeenv==1.5.0
normality==2.1.1
numpy==1.18.0
oandapyV20==0.6.3
oauth2client==4.1.3
oauthlib==3.1.0
opencage==1.2.2
openpyxl==2.5.1
packaging==20.4
page-objects==1.1.0
pandas==1.0.3
pandas-ta==0.2.23b0
paramiko==2.4.0
parse==1.8.2
parse-type==0.4.2
parsel==1.6.0
pathspec==0.8.0
pdfminer.six==20170720
phonenumbers==8.12.24
Pillow==7.2.0
platformdirs==2.0.2
pluggy==0.13.1
pre-commit==2.7.1
prometheus-client==0.8.0
prompt-toolkit==3.0.19
Protego==0.1.16
protobuf==3.13.0
psutil==5.7.2
py==1.8.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pycares==3.1.1
pycountry==18.2.23
pycparser==2.20
pycryptodome==3.9.8
PyDispatcher==2.0.5
pyee==7.0.4
PyHamcrest==1.9.0
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pypiwin32==223
pyppeteer==0.2.2
pyquery==1.4.3
pyrsistent==0.17.3
PySocks==1.7.1
pytest==6.0.1
pytest-bdd==2.18.2
python-dateutil==2.8.1
python-editor==1.0.4
python-etcd==0.4.5
pytrends==4.7.3
pytz==2021.1
pywin32==228
PyYAML==5.3.1
qrcode==7.3.1
queuelib==1.5.0
ratelim==0.1.6
redis==3.5.3
regex==2020.10.15
requests==2.22.0
requests-html==0.10.0
requests-oauth2==0.3.0
requests-oauthlib==1.3.0
responses==0.11.0
rq==1.5.1
rsa==4.6
scipy==1.5.3
Scrapy==2.4.1
selenium==3.14.1
service-identity==18.1.0
simplejson==3.17.2
six==1.13.0
soupsieve==1.9.5
SpeechRecognition==3.8.1
SQLAlchemy==1.3.19
sqlparse==0.3.1
tabula-py==2.3.0
text-unidecode==1.3
toml==0.10.1
tornado==6.1
tqdm==4.61.2
tweepy==3.8.0
Twisted==20.3.0
typed-ast==1.4.1
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.25.10
vine==1.3.0
virtualenv==20.6.0
w3lib==1.22.0
wcwidth==0.2.5
websockets==8.1
yarl==1.6.3
yfinance==0.1.54
zipp==3.5.0
zope.event==4.5.0
zope.interface==5.2.0
